Rewrote and reorganised code to be more user friendly
