+++ new_example/gauss_x_train.dat
@@ -1,5 +1,10 @@
-3.267947	-7.236809	1.296444	0.068706	0.143060	0.948665
-0.719352	1.093613	1.583695	-0.506887	-0.058200	-2.762207
-0.070236	0.331872	1.498166	-0.185071	0.909027	4.597495
-1.744593	-4.532918	1.178396	0.475965	-0.897350	-6.210051
--1.833893	8.775967	1.609641	-0.707673	0.508100	2.574929
+2.739984664270843417e+00 2.056963996093084113e+00 1.449370199931741432e+00 -3.053779629353689171e-02 -1.254985486932387595e-01 -2.115436596090754517e-01
+2.842841152453522646e+00 2.141715703691722705e+00 1.400658225287961578e+00 -4.841128120051178818e-02 2.104993966528176386e-03 1.256809210600737074e-02
+2.806513505412791876e+00 1.843464212062877472e+00 1.484429303274141132e+00 -2.982361338744663262e-02 -2.562726866701616965e-02 -1.612185714852057972e-01
+3.244362044718617888e+00 2.338777818342239634e+00 1.378995015571394589e+00 -5.034163173020086285e-02 -5.910296781699041907e-02 -2.550367860359661898e-01
+2.855990176992461826e+00 2.007022280324704422e+00 1.412969629199090527e+00 4.650496206245468417e-02 -1.769703580543564581e-01 -7.108230879215838849e-01
+2.773796425323594939e+00 2.043961449311362699e+00 1.350720539670939413e+00 1.896147714486007274e-02 -8.122253980165106135e-02 -4.298395626713700790e-01
+2.912455351211892918e+00 1.871397258102243111e+00 1.420598852656800970e+00 -3.367701439710810152e-03 8.192474949590461408e-02 -7.616131396559185085e-02
+2.777479411931308206e+00 2.239098806056040836e+00 1.432820314924655003e+00 -8.648833332330524426e-02 -1.872234091067687817e-02 1.957449975764820049e-01
+2.726994246052995230e+00 1.957450306713463206e+00 1.431005028511150634e+00 1.724083186536522943e-03 2.673949878690818141e-03 6.385965899714191041e-02
+2.625369963958456676e+00 1.824225775485190493e+00 1.396393954416408123e+00 3.168358733862255971e-02 1.640856654080490376e-02 6.172035807173763267e-02
