+++ new_example/gauss_y_train.dat
@@ -1,5 +1,10 @@
-18.753471263484045
-61.103888919772984
-50.16538974608273
-953.1789691081442
-59.36865112173409
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
