+++ new_example/hessian_diag.py
@@ -1,0 +1,8 @@
+import numpy.linalg as LA
+import numpy as np
+
+H = np.loadtxt('rigid_body_hessian.dat')
+
+eig, vec = LA.eigh(H)
+
+print(eig)
