+++ new_example/hessian_x_train.dat
@@ -1,12 +1,12 @@
-1.346036	-0.452721	1.433804	0.000000	0.000000	0.000000
-1.345836	-0.452721	1.433804	0.000000	0.000000	0.000000
-1.345936	-0.452621	1.433804	0.000000	0.000000	0.000000
-1.345936	-0.452821	1.433804	0.000000	0.000000	0.000000
-1.345936	-0.452721	1.433904	0.000000	0.000000	0.000000
-1.345936	-0.452721	1.433704	0.000000	0.000000	0.000000
-1.345936	-0.452721	1.433804	0.000100	0.000000	0.000000
-1.345936	-0.452721	1.433804	-0.000100	0.000000	0.000000
-1.345936	-0.452721	1.433804	0.000000	0.000100	0.000000
-1.345936	-0.452721	1.433804	0.000000	-0.000100	0.000000
-1.345936	-0.452721	1.433804	0.000000	0.000000	0.000100
-1.345936	-0.452721	1.433804	0.000000	0.000000	-0.000100
+1.333901205726359285e+00 -4.805825998094196416e-01 1.430511559751835637e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+1.313901205726359267e+00 -4.805825998094196416e-01 1.430511559751835637e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+1.323901205726359276e+00 -4.705825998094196327e-01 1.430511559751835637e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+1.323901205726359276e+00 -4.905825998094196505e-01 1.430511559751835637e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+1.323901205726359276e+00 -4.805825998094196416e-01 1.440511559751835646e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+1.323901205726359276e+00 -4.805825998094196416e-01 1.420511559751835629e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+1.323901205726359276e+00 -4.805825998094196416e-01 1.430511559751835637e+00 1.000000000000000021e-02 0.000000000000000000e+00 0.000000000000000000e+00
+1.323901205726359276e+00 -4.805825998094196416e-01 1.430511559751835637e+00 -1.000000000000000021e-02 0.000000000000000000e+00 0.000000000000000000e+00
+1.323901205726359276e+00 -4.805825998094196416e-01 1.430511559751835637e+00 0.000000000000000000e+00 1.000000000000000021e-02 0.000000000000000000e+00
+1.323901205726359276e+00 -4.805825998094196416e-01 1.430511559751835637e+00 0.000000000000000000e+00 -1.000000000000000021e-02 0.000000000000000000e+00
+1.323901205726359276e+00 -4.805825998094196416e-01 1.430511559751835637e+00 0.000000000000000000e+00 0.000000000000000000e+00 1.000000000000000021e-02
+1.323901205726359276e+00 -4.805825998094196416e-01 1.430511559751835637e+00 0.000000000000000000e+00 0.000000000000000000e+00 -1.000000000000000021e-02
