+++ new_example/hessian_y_train.dat
@@ -1,12 +1,12 @@
-9.457682882936211
-9.457577994351139
-9.457668664903348
-9.457592184342085
-9.45748749510257
-9.457773831322605
-9.457544499682905
-9.457716846161528
-9.4576971019068
-9.457564878881644
-9.457624696626903
-9.457636222555076
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
