+++ new_example/random_x_train.dat
@@ -1,5 +1,10 @@
-2.014535	1.924563	5.920164	-2.845335	1.247765	-3.101943
-1.342662	0.116466	3.185548	-2.829203	0.511863	0.212179
-2.410900	0.597109	0.700146	0.095185	1.036450	0.307600
-1.097465	1.383934	4.688933	-1.717843	0.487122	0.905465
-2.593124	1.661417	4.856281	1.967564	-0.604816	2.947404
+2.711055774793183382e+00 2.084123650863227262e+00 1.458275974540819364e+00 -1.186608194880876210e+00 -1.043635860333989029e-01 -1.503005472411726551e+00
+3.175635465214197151e+00 1.745915573093235418e+00 2.370311939086064967e+00 -1.440352180209427679e-01 -2.051766466055402471e-01 -6.174755657520698549e-01
+3.061479760608555090e+00 1.822025010527360322e+00 5.722008899008175220e-01 3.492372834619343891e-01 -1.535948186293354834e+00 -8.852375935959666542e-01
+2.464886132487234871e+00 2.988243164509598593e-01 4.150166306458176191e-01 -9.535984411055156151e-01 1.478536757886048658e-01 2.733929355603565980e+00
+2.220137390027927538e+00 1.904626665265815255e+00 1.649291312526938169e+00 -2.905621287191142876e+00 -2.720591159177041529e-01 1.832806459299957780e+00
+3.640900448005200296e+00 2.270607583316270439e+00 6.051857759483176835e-01 -1.104940486956365486e+00 -8.725765888275622828e-01 -1.549974059067185328e+00
+3.719909574741266489e+00 2.444300112800591229e+00 1.474158375204436044e+00 -2.423996409599842750e+00 4.477960706439128025e-01 1.875836858017217779e+00
+3.133826746871060109e+00 2.423584384867370822e+00 4.312651757610448389e-01 -2.213088955757388732e+00 -9.523237311283673234e-01 9.981092511174587401e-01
+2.815590096688220356e+00 1.591492691036073825e+00 4.695200295071959573e-01 2.835317499115826756e+00 7.058619977459112960e-02 -1.048070804377773335e+00
+2.235546584035692952e+00 2.147354434133912360e+00 1.169740679116328730e+00 -3.085327115351238714e+00 2.101586101645829263e-01 2.254964731914831155e+00
