+++ new_example/random_y_train.dat
@@ -1,5 +1,10 @@
-19.186927789416302
-15.964937234798677
-315.5363638865432
-18.516478679379514
-19.16401181818481
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
