+++ new_example/rigid_body_hessian.dat
@@ -1,0 +1,6 @@
+3.449106129793343412e-01 -5.167934294751760915e-02 1.321076843741854911e-01 6.507732973920924424e-01 -8.989576633931325444e-01 3.526062739362978982e-01
+-5.167934294751760915e-02 3.275194908107524894e-01 1.461047670069351989e-01 4.494780824541061426e-01 5.426591464199147508e-01 -1.313112027928422798e-01
+1.321076843741854911e-01 1.461047670069351989e-01 8.616003651301351596e+00 1.642223674930838317e+00 2.314676225952852218e-01 -1.638631498629953415e-02
+6.507732973920924424e-01 4.494780824541061426e-01 1.642223674930838317e+00 9.257084332485616684e+00 -7.876622296528328349e-01 6.635910400316503699e-01
+-8.989576633931325444e-01 5.426591464199147508e-01 2.314676225952852218e-01 -7.876622296528328349e-01 1.060510684801974612e+01 -2.745091430935290511e+00
+3.526062739362978982e-01 -1.313112027928422798e-01 -1.638631498629953415e-02 6.635910400316503699e-01 -2.745091430935290511e+00 8.937153957712810071e-01
