--- new_example/rigid_hessian.out
@@ -1,6 +1,0 @@
-1.085214362585111170e+00 2.149546003319156295e-01 -1.323837036221298114e+00 1.030561590025484087e+00 -2.890586216327906688e-01 1.663224597489837142e-01
-5.784610218628932765e-01 1.597249687034298371e-01 -9.519288238413229353e-01 1.382031046142812247e+00 2.405709598629028889e-01 9.674148454558859078e-02
--3.121040601660197922e+00 -1.439396927012483340e+00 1.122605678415137831e+01 1.041462851435819292e+00 -6.297554853841007549e-01 -3.025839649478113125e-01
-2.287559085634712996e+00 3.786532296586867519e-01 1.300257400846815514e+00 1.225582198739792794e+01 -1.438697601955799543e+00 1.201708681391745914e+00
-9.577411968410830534e-01 -1.444589131474325328e-01 -5.857245354750473254e-01 -8.838694194245373392e-01 1.247834576535839801e+01 -3.198953871962134876e+00
-9.401759071781042032e-02 8.188383897003725931e-02 -1.941746731539856774e-01 5.001348100244439321e-01 -2.373996285259938333e+00 5.998473483434757503e-01
