+++ new_example/sobol_x_train.dat
@@ -1,5 +1,10 @@
-2.207049	0.616951	4.858537	-1.570796	-0.523599	-1.570796
-1.685541	1.850852	1.858537	1.570796	0.523599	1.570796
-1.459721	0.925426	4.108537	-2.356194	-0.848062	2.356194
-3.406016	2.159327	1.108537	0.785398	0.252680	-0.785398
-1.720475	0.308475	2.608537	-0.785398	0.848062	0.785398
+3.406015550052440943e+00 2.159327333048420616e+00 1.106816932104180973e+00 7.853981633974482790e-01 -3.926990816987241395e-01 -7.853981633974482790e-01
+2.393880553976266867e+00 5.398318332621051541e-01 9.193169321041809727e-01 -2.159844949342982545e+00 9.817477042468114590e-02 -2.552544031041707129e+00
+2.554511662539330707e+00 1.619495499786315351e+00 5.443169321041809727e-01 -9.817477042468105708e-01 -4.908738521234052854e-01 2.552544031041707129e+00
+2.212271994022682264e+00 2.429243249679473138e+00 4.505669321041809727e-01 -1.668971097219577704e+00 7.363107781851079281e-01 -1.668971097219577704e+00
+1.915883746904498253e+00 1.214621624839736569e+00 5.911919321041809727e-01 1.472621556370214968e-01 3.681553890925539640e-01 -6.381360077604267822e-01
+3.560545736968650843e+00 1.976170103905920516e+00 1.505254432104180973e+00 -1.742602175038088452e+00 -4.049709280018092272e-01 -1.104466167277661448e+00
+3.086487452535218612e+00 1.455617978974604876e+00 5.208794321041809727e-01 -2.871612034921920298e+00 2.577087723647877304e-01 -3.681553890925539640e-01
+2.793139165235233978e+00 2.149687478883025804e+00 1.458379432104180973e+00 2.822524649709579947e+00 5.522330836388307240e-01 2.969786805346601000e+00
+2.963713149238814459e+00 1.768913239349933608e+00 7.200981821041809727e-01 -1.951223562190535610e+00 -2.270291566070747891e-01 2.785709110800324240e+00
+1.781899918243328163e+00 2.193066822627302237e+00 4.388481821041809727e-01 -1.067650628368406185e+00 1.165825398793087775e-01 -2.221204180858408428e+00
