+++ new_example/sobol_y_train.dat
@@ -1,5 +1,10 @@
-19.048312641099756
-14.80566849775051
-16.75062254572563
-34.86132939887909
-26.251912103502427
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
+3.644020611524068727e+02
